first step for create whitelabel

+++ Modules/Groups/Database/Migrations/2018_12_14_132158_create_groups_table.php
@@ -26,7 +26,7 @@ class CreateGroupsTable extends Migration
             $table->timestamps();
         });
 
-        if (!Schema::hasColumn('wishes', 'group_id')) {
+        if (Schema::hasTable('wishes') && !Schema::hasColumn('wishes', 'group_id')) {
             Schema::table('wishes', function (Blueprint $table) {
                 $table->integer('group_id')->after('created_by')->nullable()->unsigned()->index();
                 $table->foreign('group_id')->references('id')->on('groups')->onDelete('cascade');
